backend/requirements.txt
fastapi
uvicorn
pydantic
torch
torchaudio
librosa
numpy
demucs
openai-whisper
# whisperx # Optional but recommended for better alignment
crepe
python-multipart
soundfile
ffmpeg-python
